requirements.txt
Flask==1.1.2
flask_migrate==2.5.3
flask_moment==0.10.0
flask_sqlalchemy==2.4.4
python-dotenv==0.15.0
